requirements.txt
GitPython>=3.1.43
matplotlib>=3.8.4
numpy>=1.24.3
pywin32>=305.1
sqlalchemy>=1.4
